requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
groq==0.3.1
python-dotenv==0.20.0
flask-cors==3.1.1
